requirements.txt
dnspython==1.11.0
lxml==3.2.1
python-dateutil==2.1
six==1.3.0
sleekxmpp==1.1.11
wsgiref==0.1.2
